requirements.txt
kopf
azure-mgmt-dns
azure-identity